+++ sdk/dynamicBrcode/rule/index.js
@@ -1,0 +1,3 @@
+const rule = require('./rule.js');
+
+exports.Rule = rule.Rule;